requirements.txt
wandb~=0.13.1
yaml~=0.2.5
pyyaml~=6.0
tqdm~=4.64.0
dotmap~=1.3.30
pillow~=9.0.1
torchvision~=0.13.0
numpy~=1.22.4
ftfy~=6.1.1
regex~=2022.3.15
pandas~=1.4.2
scikit-learn~=1.0.2
opencv-python~=4.6.0.66
setuptools~=61.2.0
matplotlib~=3.5.1
seaborn~=0.11.2